task: renaming a model and adding home
