+++ app/actions/home/home.go
@@ -1,0 +1,18 @@
+package home
+
+import (
+	"mjm/app/render"
+	"net/http"
+
+	"github.com/gobuffalo/buffalo"
+)
+
+var (
+	// r is a buffalo/render Engine that will be used by actions
+	// on this package to render render HTML or any other formats.
+	r = render.Engine
+)
+
+func Index(c buffalo.Context) error {
+	return c.Render(http.StatusOK, r.HTML("home/index.plush.html"))
+}
