+++ app/actions/requirement_type/requirement_type.go
@@ -1,4 +1,4 @@
-package requerimenttype
+package requirement_type
 
 import (
 	"fmt"
