+++ app/actions/requirement_type/requirement_type_test.go
@@ -1,4 +1,4 @@
-package requerimenttype_test
+package requirement_type_test
 
 import (
 	"mjm/app"
